Small typos in pluto.jl
I absolutely loved reading this post! It seems like this is a missing part of a researcher / programmer's education that should be more common to teach - because (at most) even in CS departments we learn about, for example, Big O Notation and it's only in rare niche courses that we learn all these details.

Anyway I really enjoyed reading this, and I'll be sharing it with my colleagues! It has me thinking that (along with my point about about needing more teaching here) it would be neat to actually analyze some code in the wild, compare techniques to do calculations, and maybe have an automated way to then make those suggestions.